requirements/base.txt
Pillow

django-ckeditor
social-auth-app-django
rq
django-rq
rq-scheduler